sep file: keyboard event translation
