+++ src/lib/mod.rs
@@ -6,3 +6,4 @@ extern crate serde_bytes;
 //extern crate candid_derive;
 
 pub mod types;
+pub mod keyboard;
